wip: simplify payload image component
