+++ packages/ui/components/primitives/RichText/index.tsx
@@ -2,7 +2,7 @@ import s from '@refract-ui/sc';
 import type React from 'react';
 import type { DefaultTheme } from 'styled-components';
 
-import type { ResponsivePayloadWrapperProps } from '../ResponsivePayloadImage/ResponsivePayloadImage.client';
+import type { ResponsivePayloadWrapperProps } from '../ResponsivePayloadImage';
 
 import serializeText from './utils/serializeText';
